Resolve VRDRUM-161 "Main system add tag to lesson category add some validation add lesson level name in"

Closes VRDRUM-161

See merge request development-team13/vr-drumming-academy/main_system!9

+++ app/Http/Controllers/Admin/LessonWithLessonCoachController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\UpdateLessonRequest;
 use App\Models\CoachDetail;
 use App\Models\LessonCoach;
 use App\Models\Lesson;
+use App\Models\LessonCategory;
 use App\Models\LessonLevel;
 use Gate;
 use Validator;
@@ -86,11 +87,17 @@ class LessonWithLessonCoachController extends Controller
     {
         abort_if(Gate::denies('lesson_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        $lesson_levels = LessonLevel::pluck('level', 'id')->prepend(trans('global.pleaseSelect'), '');
+        $lesson_levels = [ '' => trans('global.pleaseSelect') ];
+
+        $all_lesson_level = LessonLevel::all();
+        foreach ($all_lesson_level as $index => $value){
+            $lesson_category = LessonCategory::find($value->lesson_category_id);
+            $lesson_levels[$value->id] = $lesson_category->name . ' - ' . $value->level;
+        }
 
         $coachs = CoachDetail::pluck('coach_efk', 'coach_efk');
 
-        if($errors != null) return view('admin.lessons.create', compact('lesson_levels', 'coachs', 'errors'));
+        if($errors != null) return view('admin.lessons.create', compact('lesson_levels', 'coachs'))->withErrors($errors);
 
         return view('admin.lessons.create', compact('lesson_levels', 'coachs'));
     }
@@ -100,14 +107,14 @@ class LessonWithLessonCoachController extends Controller
         $request_data = $request->all();
 
         // ------------------------------ validation ------------------------------
-        $validated = Validator::make([],[]);
+        $errors = [];
 
         if ($request_data['lesson_level_id'] == null){
-            $validated->getMessageBag()->add('lesson_level', trans('validation.lesson_level_required'));
+            $errors['lesson_level'] = trans('validation.lesson_level_required');
         }
 
-        if($validated->errors()->count() > 0){
-            return $this->create($validated->errors());
+        if(count($errors) > 0){
+            return $this->create($errors);
         }
 
         // ------------------------------ data assign ------------------------------
@@ -136,9 +143,15 @@ class LessonWithLessonCoachController extends Controller
     {
         abort_if(Gate::denies('lesson_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        $lesson_levels = LessonLevel::pluck('level', 'id')->prepend(trans('global.pleaseSelect'), '');
-
         $lesson->load('lesson_level', 'created_by');
+
+        $lesson_levels = [ '' => trans('global.pleaseSelect') ];
+
+        $all_lesson_level = LessonLevel::all();
+        foreach ($all_lesson_level as $index => $value){
+            $lesson_category = LessonCategory::find($value->lesson_category_id);
+            $lesson_levels[$value->id] = $lesson_category->name . ' - ' . $value->level;
+        }
 
         $coachs = CoachDetail::pluck('coach_efk', 'coach_efk');
 
@@ -149,7 +162,7 @@ class LessonWithLessonCoachController extends Controller
             $old_coachs_efk[$value->coach_efk] = $value->coach_efk;
         }
         
-        if($errors != null) return view('admin.lessons.edit', compact('lesson', 'lesson_levels', 'coachs', 'old_coachs_efk', 'errors'));
+        if($errors != null) return view('admin.lessons.edit', compact('lesson', 'lesson_levels', 'coachs', 'old_coachs_efk'))->withErrors($errors);
 
         return view('admin.lessons.edit', compact('lesson', 'lesson_levels', 'coachs', 'old_coachs_efk'));
     }
@@ -159,14 +172,14 @@ class LessonWithLessonCoachController extends Controller
         $request_data = $request->all();
 
         // ------------------------------ validation ------------------------------
-        $validated = Validator::make([],[]);
+        $errors = [];
 
         if ($request_data['lesson_level_id'] == null){
-            $validated->getMessageBag()->add('lesson_level', trans('validation.lesson_level_required'));
+            $errors['lesson_level'] = trans('validation.lesson_level_required');
         }
 
-        if($validated->errors()->count() > 0){
-            return $this->edit($lesson, $validated->errors());
+        if(count($errors) > 0){
+            return $this->edit($lesson, $errors);
         }
 
         // ------------------------------ data assign ------------------------------
@@ -233,6 +246,13 @@ class LessonWithLessonCoachController extends Controller
     public function destroy(Lesson $lesson)
     {
         abort_if(Gate::denies('lesson_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        
+        // ------------------------------ validation ------------------------------
+        $last_lesson = Lesson::where('lesson_level_id', $lesson->lesson_level_id)->orderBy('no_of_class', 'DESC')->get()->first();
+
+        if($last_lesson != null && $lesson->id != $last_lesson->id){
+            return back()->withErrors(['lesson_level' => trans('validation.lesson_level_delete_last')]);
+        }
 
         $lesson->delete();
 
@@ -241,7 +261,9 @@ class LessonWithLessonCoachController extends Controller
 
     public function massDestroy(MassDestroyLessonRequest $request)
     {
-        Lesson::whereIn('id', request('ids'))->delete();
+        // Lesson::whereIn('id', request('ids'))->delete();
+        // $lesson = Lesson::whereIn('id', request('ids'))->get();
+        // return response()->json($lesson);
 
         return response(null, Response::HTTP_NO_CONTENT);
     }
